Bug Fix
Tried a few things to get the moving platforms to collide with walls.
Only thing that worked was making the moving platforms have Collision
Masks at all times and not act as one way Platforms. Bug seems to have
something to do with the collision masks not spawning if the oPlayer is
below them.